css/cssom/computed-style-004.html

<!DOCTYPE html>
<title>CSS Test: getComputedStyle - resolved width in iframe dynamic display</title>
<link rel="author" title="Rune Lillesveen" href="mailto:futhark@chromium.org" />
<link rel="help" href="https://drafts.csswg.org/cssom/#resolved-values" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="frm" width="100" style="display:none"></iframe>
<script>
  const frmDoc = frm.contentWindow.document;
  frmDoc.open();
  frmDoc.write('<body style="margin:0"><div style="width:100%"></div>');
  frmDoc.close();

  test(() => {
    document.body.offsetWidth; // Make sure we layout the top document.
    assert_equals(frm.contentWindow.getComputedStyle(frmDoc.querySelector("div")).width, "100%");
  }, "Check that a percent width in a display:none iframe is the computed value of width for getComputedStyle.");

  test(() => {
    frm.style.display = "inline";
    assert_equals(frm.contentWindow.getComputedStyle(frmDoc.querySelector("div")).width, "100px");
  }, "Check that a percent width in an iframe is the resolved width when the iframe is displayed.");
</script>
